Move formatDate to common module for reusability
